chore: move types to the test package
