+++ modules/rabbitmq/rabbitmq_test.go
@@ -14,17 +14,17 @@ func TestRunContainer_withAllSettings(t *testing.T) {
 	rabbitmqContainer, err := rabbitmq.RunContainer(ctx,
 		testcontainers.WithImage("rabbitmq:3.7.25-management-alpine"),
 		// addVirtualHosts {
-		testcontainers.WithStartupCommand(rabbitmq.VirtualHost{Name: "vhost1"}),
-		testcontainers.WithStartupCommand(rabbitmq.VirtualHostLimit{VHost: "vhost1", Name: "max-connections", Value: 1}),
-		testcontainers.WithStartupCommand(rabbitmq.VirtualHost{Name: "vhost2", Tracing: true}),
+		testcontainers.WithStartupCommand(VirtualHost{Name: "vhost1"}),
+		testcontainers.WithStartupCommand(VirtualHostLimit{VHost: "vhost1", Name: "max-connections", Value: 1}),
+		testcontainers.WithStartupCommand(VirtualHost{Name: "vhost2", Tracing: true}),
 		// }
 		// addExchanges {
-		testcontainers.WithStartupCommand(rabbitmq.Exchange{Name: "direct-exchange", Type: "direct"}),
-		testcontainers.WithStartupCommand(rabbitmq.Exchange{
+		testcontainers.WithStartupCommand(Exchange{Name: "direct-exchange", Type: "direct"}),
+		testcontainers.WithStartupCommand(Exchange{
 			Name: "topic-exchange",
 			Type: "topic",
 		}),
-		testcontainers.WithStartupCommand(rabbitmq.Exchange{
+		testcontainers.WithStartupCommand(Exchange{
 			VHost:      "vhost1",
 			Name:       "topic-exchange-2",
 			Type:       "topic",
@@ -33,12 +33,12 @@ func TestRunContainer_withAllSettings(t *testing.T) {
 			Durable:    true,
 			Args:       map[string]interface{}{},
 		}),
-		testcontainers.WithStartupCommand(rabbitmq.Exchange{
+		testcontainers.WithStartupCommand(Exchange{
 			VHost: "vhost2",
 			Name:  "topic-exchange-3",
 			Type:  "topic",
 		}),
-		testcontainers.WithStartupCommand(rabbitmq.Exchange{
+		testcontainers.WithStartupCommand(Exchange{
 			Name:       "topic-exchange-4",
 			Type:       "topic",
 			AutoDelete: false,
@@ -48,26 +48,26 @@ func TestRunContainer_withAllSettings(t *testing.T) {
 		}),
 		// }
 		// addQueues {
-		testcontainers.WithStartupCommand(rabbitmq.Queue{Name: "queue1"}),
-		testcontainers.WithStartupCommand(rabbitmq.Queue{
+		testcontainers.WithStartupCommand(Queue{Name: "queue1"}),
+		testcontainers.WithStartupCommand(Queue{
 			Name:       "queue2",
 			AutoDelete: true,
 			Durable:    false,
 			Args:       map[string]interface{}{"x-message-ttl": 1000},
 		}),
-		testcontainers.WithStartupCommand(rabbitmq.Queue{
+		testcontainers.WithStartupCommand(Queue{
 			VHost:      "vhost1",
 			Name:       "queue3",
 			AutoDelete: true,
 			Durable:    false,
 			Args:       map[string]interface{}{"x-message-ttl": 1000},
 		}),
-		testcontainers.WithStartupCommand(rabbitmq.Queue{VHost: "vhost2", Name: "queue4"}),
+		testcontainers.WithStartupCommand(Queue{VHost: "vhost2", Name: "queue4"}),
 		// }
 		// addBindings {
-		testcontainers.WithStartupCommand(rabbitmq.NewBinding("direct-exchange", "queue1")),
-		testcontainers.WithStartupCommand(rabbitmq.NewBindingWithVHost("vhost1", "topic-exchange-2", "queue3")),
-		testcontainers.WithStartupCommand(rabbitmq.Binding{
+		testcontainers.WithStartupCommand(NewBinding("direct-exchange", "queue1")),
+		testcontainers.WithStartupCommand(NewBindingWithVHost("vhost1", "topic-exchange-2", "queue3")),
+		testcontainers.WithStartupCommand(Binding{
 			VHost:           "vhost2",
 			Source:          "topic-exchange-3",
 			Destination:     "queue4",
@@ -77,33 +77,33 @@ func TestRunContainer_withAllSettings(t *testing.T) {
 		}),
 		// }
 		// addUsers {
-		testcontainers.WithStartupCommand(rabbitmq.User{
+		testcontainers.WithStartupCommand(User{
 			Name:     "user1",
 			Password: "password1",
 		}),
-		testcontainers.WithStartupCommand(rabbitmq.User{
+		testcontainers.WithStartupCommand(User{
 			Name:     "user2",
 			Password: "password2",
 			Tags:     []string{"administrator"},
 		}),
 		// }
 		// addPermissions {
-		testcontainers.WithStartupCommand(rabbitmq.NewPermission("vhost1", "user1", ".*", ".*", ".*")),
+		testcontainers.WithStartupCommand(NewPermission("vhost1", "user1", ".*", ".*", ".*")),
 		// }
 		// addPolicies {
-		testcontainers.WithStartupCommand(rabbitmq.Policy{
+		testcontainers.WithStartupCommand(Policy{
 			Name:       "max length policy",
 			Pattern:    "^dog",
 			Definition: map[string]interface{}{"max-length": 1},
 			Priority:   1,
 			ApplyTo:    "queues",
 		}),
-		testcontainers.WithStartupCommand(rabbitmq.Policy{
+		testcontainers.WithStartupCommand(Policy{
 			Name:       "alternate exchange policy",
 			Pattern:    "^direct-exchange",
 			Definition: map[string]interface{}{"alternate-exchange": "amq.direct"},
 		}),
-		testcontainers.WithStartupCommand(rabbitmq.Policy{
+		testcontainers.WithStartupCommand(Policy{
 			VHost:   "vhost2",
 			Name:    "ha-all",
 			Pattern: ".*",
@@ -112,7 +112,7 @@ func TestRunContainer_withAllSettings(t *testing.T) {
 				"ha-sync-mode": "automatic",
 			},
 		}),
-		testcontainers.WithStartupCommand(rabbitmq.OperatorPolicy{
+		testcontainers.WithStartupCommand(OperatorPolicy{
 			Name:       "operator policy 1",
 			Pattern:    "^queue1",
 			Definition: map[string]interface{}{"message-ttl": 1000},
@@ -121,7 +121,7 @@ func TestRunContainer_withAllSettings(t *testing.T) {
 		}),
 		// }
 		// enablePlugins {
-		testcontainers.WithStartupCommand(rabbitmq.Plugin("rabbitmq_shovel"), rabbitmq.Plugin("rabbitmq_random_exchange")),
+		testcontainers.WithStartupCommand(Plugin("rabbitmq_shovel"), Plugin("rabbitmq_random_exchange")),
 		// }
 	)
 	if err != nil {
